updated `sort` column type from `INTEGER` to `NUMERIC` in product images and videos migration files to support more precise sorting values

+++ migrations/20260215164005_create_product_videos_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE product_videos
     url        TEXT        NOT NULL,
     alt        TEXT        NOT NULL,
     is_main    BOOLEAN     NOT NULL DEFAULT false,
-    sort       INTEGER     NOT NULL DEFAULT 0,
+    sort       NUMERIC     NOT NULL DEFAULT 0,
     deleted_at TIMESTAMPTZ,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
